// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Event: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly tribe: CodecTypes['pg/text@1']['output'];
      readonly groomName: CodecTypes['pg/text@1']['output'];
      readonly eventDate: CodecTypes['pg/date-string@1']['output'];
      readonly venue: CodecTypes['pg/text@1']['output'] | null;
      readonly imageUrl: CodecTypes['pg/text@1']['output'];
      readonly galleryUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly shortCode: CodecTypes['pg/text@1']['output'];
      readonly clickCount: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: TimestampString<3>;
      readonly updatedAt: TimestampString<3>;
    };
    readonly EventRequest: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly phone: CodecTypes['pg/text@1']['output'];
      readonly preferredDate: CodecTypes['pg/date-string@1']['output'] | null;
      readonly venue: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'NEW' | 'CONTACTED' | 'CLOSED' | 'PUBLISHED';
      readonly createdAt: TimestampString<3>;
    };
  };
};